powerstats: Fix unnecessary odpm reading
When getEnergyConsumed is called with empty customer id or a list of
customer ids, the odpm reading (via reading a iio sysfs) will be
triggered multiple times and causing unnecessary odpm re-sampling in the
PMIC. Instead refresh the odpm reading first and passing to each
customer to add up the power to avoid this.

Also refacctor the dump() function to make sure odpm is refresh only
once when dump() is called.

Bug: 402289949
Test: dumpsys android.hardware.power.stats.IPowerStats/default
Flag: EXEMPT bugfix
Change-Id: I8749da6fc81075c88ac673f10c1edce5d5495ad7