Add support for powerhint.json flagging and add a flag provider
This patch adds a new class called FlagProvider which tracks flags
and allows for custom override flag values to be set for testing.

It also tracks the string names of flags and allows them to be
associated with the getter for that flag, and this patch uses that
functionality to then add support for flagging the powerhint.json file.

Bug: 438441514
Test: atest libperfmgr_test
Test: atest libadpf_test

Change-Id: I77882eba6fc9468dfb928cefdb5b128e4fecda52