Allow debug powerhint to fall back to the default one
Setting "persist.vendor.powerhal.config.debug.usefallback" to 1 allows
the debug powerhint to fall back to the standard powerhint.json if the
config fails to load, avoiding a powerhal crash. An error will be
written to logcat in this case, which users can check for to determine
if the load failed.

Bug: 434755015
Test: atest libperfmgr_test
Change-Id: Ib3297051e1e8a4b8e7211fc24dc964ef411d4c2e