libpixelusb: set invalid usb state log level to debug
On platforms where the UDC device is not always available in
host mode, a timing issue can prevent the UDC epoll setup
from being properly cleared when the platform is booted with
host mode enabled. This results in the UDC file descriptor
continuously reporting an invalid state and creating log spam.
This behavior is benign as it doesn't have any functional impact.
This change moves the relevant logs to a lower debug level to
avoid log spam.

Bug: 436981349
Test: trigger usb state update
Flag: EXEMPT trivial bug fix
Change-Id: If04062150069a0a5cfd6d0a6b7a2577cbcea2d7c