Monitorffs: disable some logs in USB host mode
When the device is in USB host mode, it's normal for
Monitorffs to generate error logs from access failures.
These logs are expected and can be safely ignored. In
addition, epoll_wait() would receive an error "EINTR"
when the device entered suspend. It is a normal case
so that the log could be also ignored.

Bug: 434043948
Test: logs aren't print
Change-Id: I9114236c12f058ae8ac8703a04a83a9f213ab2a5